Added an option to the SchemaPrinter to force printing of the schema definition.
Currently the schema printer won't print the schema definition if the query, mutation, and
subscription types use the default names. Some other graphql libraries (python graphql for instance)
require the schema definition to always be present. So this option just forces printing the
schema definition for those consumers that need/want it.