Enum default values should use EnumValuesProvider (#2150)
If you create a schema with SchemaGenerator with an EnumValuesProvider, you
would expect that your DataFetchers can call getArgument and get enum values as
returned by the EnumValuesProvider. This works fine when the value is explicitly
provided in the operation. But for default values, DataFetchers will just
receive a String with the name of the value!  This PR improves SchemaGenerator
to resolve enum values in default values (included nested in input objects,
lists, and non-null) using EnumValuesProvider.